+++ app/src/main/java/woowacourse/shopping/database/product/RecentViewedDBContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.database.recentviewed
+package woowacourse.shopping.database.product
 
 object RecentViewedDBContract {
